@@ -1,5 +1,6 @@
 ﻿@page "/application/{Id:guid}"
-@using GdsApi;
+
+@using global::GdsApi
 @inject IApplicationsGdsApi ApplicationsGdsApi
 
 <PageTitle>Application Details</PageTitle>
