@@ -1,5 +1,5 @@
 ﻿@page "/applications"
-@using GdsApi;
+@using global::GdsApi
 @inject IApplicationsGdsApi ApplicationsGdsApi
 
 <PageTitle>Applications</PageTitle>
