@@ -1,5 +1,5 @@
 ﻿@page "/certificate-group/{Id}"
-@using GdsApi;
+@using global::GdsApi
 @inject ICertificateGroupsGdsApi CertificateGroupsGdsApi
 
 <PageTitle>CertificateGroup Details</PageTitle>
