@@ -1,5 +1,5 @@
 ﻿@page "/certificate-groups"
-@using GdsApi;
+@using global::GdsApi
 @inject ICertificateGroupsGdsApi CertificateGroupsGdsApi
 
 <PageTitle>Certificate Groups</PageTitle>
